Broke out the demoTimer function from main.js and remove unnecessary require of generateDemoBoard. #28
